splitng to train test split

+++ MachineLearning_Ds/CountryData.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.ML.Data;
+using System.Collections.Generic;
 
 namespace NullValues
 {
@@ -22,22 +23,23 @@ namespace NullValues
     {
 
         [LoadColumn(0)]
-        public int Age { get; set; }
+        public float Age { get; set; }
 
         [LoadColumn(1)]
-        public int Salary { get; set; }
+        public float Salary { get; set; }
 
         [LoadColumn(2)]
-        public int purchasedList { get; set; }
+        public float purchasedList { get; set; }
 
         [LoadColumn(3)]
-        public int Country_France { get; set; }
+        public float Country_France { get; set; }
 
         [LoadColumn(4)]
-        public int Country_Spain { get; set; }
+        public float Country_Spain { get; set; }
 
         [LoadColumn(5)]
-        public int Country_Germany { get; set; }
+        public float Country_Germany { get; set; }
     }
+
 
 }
